requirements.txt
# requirements.txt
Flask
gunicorn
requests
google-cloud-storage